[Arm64] set m_maskingIsFree for aarch64 - bug #2052
disable Coherent Control Flow optimization for aarch64 because of a bug in LLVM
aarch64 back-end that reduces the efficiency of simplifyCFG.  Branches added by
CCF can only be removed after the back-end formed fused-multiply-adds.  This
reduces the quality of code as most of scalar optimizations will not apply.

FIXME: Consider turning this optimization back on after
https://reviews.llvm.org/D100963 gets committed to LLVM-13.