Fix limit of enqueued kernel launches
The limit of kernel launches is set to a low value - 100.
This commit rises this limit to 100000. It does not increase
the memory usage by more than 4kb, so should do no harm.
This commit also introduces the ability to set a smaller limit
by using environmental variable (ISPCRT_MAX_KERNEL_LAUNCHES).