tests: fix 2514.ispc test adding --nowrap
By default, output of ISPC is wrapped to some fixed width. When ispc
test suite is placed in the relatively short or long path, output is
wrapped in a way that can disjoint the search pattern to different
lines whereas CHECK is searching for substring on the same line. It may
cause fails specific to build/test environment.