Destroy TargetMachine in Target destructor.
This is a follow up on the review of PR2575 and fixes #2578.

TargetMachine really belongs to the Target and needs to be destroyed in
its destructor.

The cases when it escapes Target lifetime are not legitimate. For the
dispatch module we create a new Target anyway and it creates a
TargetMachine as well - and this one needs to be used.

This should make the design a little cleaner and reduce memory footprint
for multi-target compilation because of earlier release of
TargetMachine.