parser: fix deallocation on parser error
Intially, it was deallocated in yyerror function in eaa5d755403d0.
It was wrong because yylval is union, so stringVal is not null for many
different types, e.g., in 2599.ispc it is small integer value on the
moment of calling yyerror.

Bison has the %destructor directive to define code that is called when a
symbol is automatically discarded, e.g., during the error recovery.
See manual for details:
https://www.gnu.org/software/bison/manual/html_node/Destructor-Decl.html

This change tag all tokens that semantically returns heap-allocated
string as <stringVal>. The destructor directive is also added for
<stringVal>. It results in yydestruct function filled with code to
deallocate corresponding to <stringVal> tokens and types. It happend
inside the body of yydestruct function in switch like this:
  switch (yytype)
    {
    case 15: /* TOKEN_FLOAT16_CONSTANT  */
            { lCleanUpyylvalStringVal(); }
        break;

    case 20: /* TOKEN_IDENTIFIER  */
            { lCleanUpyylvalStringVal(); }
        break;
    ...
    }