Fix bool swizzles
StoreInst expects value type as third argument not type of pointer to
value. Otherwise, the code inside casting bool is meaningless because
PointerType can't be bool. Moreover, other use cases shows that the third
argument is just type of the value passed as the first argument.
I don't know why argument name is so confusing here.