Fix return type of lParseOperator
When bison is used in -y mode that emulates POSIX Yacc, tokens are
defined as enums (under YYTOKENTYPE ifdef) or as int via macro
definitions. Defining return type as yytokentype causes compile error:
invalid conversion from 'int' to 'yytokentype'. To avoid it, return int
as we do with lParseInteger and lParseFP.