opt: add ReplaceHalfMaskedMemOps pass
It traverse bitcode for masked stores that have the turned-off part
half and the turned-on first part. We can safely replace them with
narrow unmasked stores and loads with the following shuffle with the
passthrough value. This can help the back-end to generate better code
(no extra spills, assigning narrow registers).