Fix overloaded LLVM intrinsics
We need to produce correct function declarations for overloaded LLVM
intrinsics. ISPC used all arguments to mangling in extra suffixes (which
is wrong). We need to provide only some arguments. At the moment, there
is only a short list of intrinsics with the manual mapping of which
arguments should be mangled. It is need to be done in general later (for
all current and future intrinsics). But at the moment, it is OK for a
limited set of intrinsics that is to be used in common-target.