dispatcher: remove dependency on C abort
Replace the dependency on the C abort function with a custom
implementation (__terminate_now), which directly invokes the ud2
instruction.

This change introduces a breaking change for compiled user programs.
Specifically, when no ISA is detected, the behavior changes from
triggering an abort to executing an illegal instruction, meaning a
different signal is raised (SIGABRT → SIGILL). User code that installs
signal handlers for SIGABRT and processes them in any way will break and
require modifications.

ISPC does not appear to guarantee any specific behavior for this case,
so this change is likely acceptable. However, it is beneficial to
document this behavior explicitly.