generic.ispc: implement masked_store_blend via @llvm.ispc.blend_store
It resolves infinite recursion in __masked_store_blend that happens
under -O0. The problem here is that ISPC generates __pseudo_masked_store
even if it knows that this is unmasked region with mask being all_on.
Later __pseudo_masked_store is substituted with __masked_store_blend
resulting in the real recursion. In reality, argument of this call
should be propagated resulting into simple load (and this happens for
-O2 pipeline).