Fix the crash of #3388
This PR fix the crash of #3388 in a naive way:
ISPC crashes because `exprs[0]` is null hence the segmentation fault. If we add a check then this does not cause any segfault on the example. That being said, this is a naive fix because I do not know if having a `nullptr` in `exprs` is actually normal. If not, then another more complex fix is needed.