Support --opt=disable-zmm for generic targets
Supported targets:
- generic-i32x16
- generic-i1x16
- generic-i1x32
- generic-i1x64

Also report warnings that --opt=disable-zmm is unsupported or erroneous
for targets that support avx512. It was implemented only for x16 wide
targets, but not every program could be compiled because builtins use
LLVM intrinsics that require zmm registers. For x32 and x64 targets, it
was not implemented at all.

On the contrary, builtins of generic targets do not use target specific
builtins so we can use them with --opt=disable-zmm and the proper CPUs,
e.g., --opt=disable-zmm --target=generic-i1x16 --cpu=skx