benchmarks: support __fp16 testing
Test during configuration step of the build process whether the used
compiler supports __fp16 type. To do that, the simple program is
compiled. If the compiler supports it then both C++ and ISPC code can
use macro-definition HAS_FP16 to guard code that use __fp16 and float16
types.

It is worth noting that ISPC needs to guard float16 functionality that
is exported to C/C++ world, because float16 is matched to __fp16 in
function or struct declarations inside the header file.