[LLVM21][NFC] Update LIT test to account for improvements in SROA.
SROA was recently updated to optimize stores that fill different parts of same
alloca - https://github.com/llvm/llvm-project/pull/152793.

This affects generated IR/ASM by ISPC for a LIT test that was used to validate
if 2 versions of source generate identical ASM. This PR updates the test, and
creates a duplicate for LLVM_21_0+ builds.